requirements.txt
blinker==1.4
Click==7.0
Flask==1.0.2
Flask-Mail==0.9.1
gerencianet==1.0.10
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.4.1
pkg-resources==0.0.0
PyJWT==1.7.1
PyMySQL==0.8.1
reportlab==3.5.13
requests==2.9.1
six==1.9.0
Werkzeug==0.14.1
